Visual fixes & IK improvements
Reduced the size of everything. Added a cube as reference to see which side is front.
Trying out just removing the unnecessary DOF in the shoulder to see if it helps accuracy in angles. Rest position for joint-level is now elbow bent 90 degrees, so that the arm can be segmented more accurately.